:recycle: HangongsuException -> BaseException 이름 변경 및 로깅 메시징 필드 추가
